refactor: make all task tab as a component
